+++ php/classes/controllers/class-admin-controller.php
@@ -145,8 +145,6 @@ class Admin_Controller extends Controller {
 			add_action( 'admin_init', array( $this, 'ignore_importing_existing_podcasts' ) );
 			add_action( 'admin_init', array( $this, 'start_importing_existing_podcasts' ) );
 
-			add_action( 'current_screen', array( $this, 'check_existing_podcasts' ) );
-
 			// Show upgrade screen
 			add_action( 'current_screen', array( $this, 'show_upgrade_screen' ), 12 );
 
@@ -1488,39 +1486,7 @@ HTML;
 		$castos_handler->delete_podcast( $post );
 	}
 
-	/**
-	 * Check if there are existing podcasts to be uploaded to Seriously Simple Hosting
-	 */
-	public function check_existing_podcasts() {
-		/**
-		 * Only trigger this if we're connected to Seriously Simple Hosting
-		 */
-		if ( ! ssp_is_connected_to_podcastmotor() ) {
-			return;
-		}
 
-		/**
-		 * Only show this notice on the All Episodes page
-		 */
-		$current_screen = get_current_screen();
-		if ( 'edit-podcast' !== $current_screen->id ) {
-			return;
-		}
-
-		/**
-		 * Only trigger this if the ss_podcasting_podmotor_import_podcasts option hasn't been set
-		 */
-		$ss_podcasting_podmotor_import_podcasts = get_option( 'ss_podcasting_podmotor_import_podcasts', '' );
-		if ( ! empty( $ss_podcasting_podmotor_import_podcasts ) ) {
-			return;
-		}
-
-		// check if there is at least one podcast to import
-		$podcast_query = ssp_get_existing_podcasts();
-		if ( $podcast_query->have_posts() ) {
-			add_action( 'admin_notices', array( $this, 'existing_podcasts_notice' ) );
-		}
-	}
 
 	/**
 	 * Setup podcast import
@@ -1546,31 +1512,7 @@ HTML;
 		}
 	}
 
-	/**
-	 * Admin notices
-	 * Ideally this should be moved to it's own controller
-	 */
 
-	/**
-	 * Show 'existing podcast' notice
-	 */
-	public function existing_podcasts_notice() {
-		$podcast_import_url = add_query_arg( array(
-			'post_type' => $this->token,
-			'page'      => 'podcast_settings',
-			'tab'       => 'import'
-		) );
-		$ignore_message_url = add_query_arg( array( 'podcast_import_action' => 'ignore' ) );
-		$message            = '';
-		$message            .= '<p>You\'ve connected to your Castos account and you have existing podcasts that can be imported.</p>';
-		$message            .= '<p>You can <a href="' . $podcast_import_url . '">import your existing podcasts to Castos.</a></p>';
-		$message            .= '<p>Alternatively you can <a href="' . $ignore_message_url . '">dismiss this message.</a></p>';
-		?>
-		<div class="notice notice-info">
-			<p><?php _e( $message, 'ssp' ); ?></p>
-		</div>
-		<?php
-	}
 
 	/**
 	 * Show 'importing podcasts' notice
