+++ php/classes/controllers/class-admin-notifications-controller.php
@@ -1,0 +1,90 @@
+<?php
+
+namespace SeriouslySimplePodcasting\Controllers;
+
+// Exit if accessed directly.
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+/**
+ * Main plugin class
+ *
+ * @author      Hugh Lashbrooke
+ * @category    Class
+ * @package     SeriouslySimplePodcasting/Controllers
+ * @since       1.0
+ */
+class Admin_Notifications_Controller extends Controller {
+
+	/**
+	 * Admin_Controller constructor.
+	 *
+	 * @param $file string main plugin file
+	 * @param $version string plugin version
+	 */
+	public function __construct( $file, $version ) {
+		parent::__construct( $file, $version );
+		$this->bootstrap();
+	}
+
+	public function bootstrap(){
+		add_action( 'current_screen', array( $this, 'check_existing_podcasts' ) );
+	}
+
+	/**
+	 * Check if there are existing podcasts to be uploaded to Seriously Simple Hosting
+	 */
+	public function check_existing_podcasts() {
+		/**
+		 * Only trigger this if we're connected to Seriously Simple Hosting
+		 */
+		if ( ! ssp_is_connected_to_podcastmotor() ) {
+			return;
+		}
+
+		/**
+		 * Only show this notice on the All Episodes page
+		 */
+		$current_screen = get_current_screen();
+		if ( 'edit-podcast' !== $current_screen->id ) {
+			return;
+		}
+
+		/**
+		 * Only trigger this if the ss_podcasting_podmotor_import_podcasts option hasn't been set
+		 */
+		$ss_podcasting_podmotor_import_podcasts = get_option( 'ss_podcasting_podmotor_import_podcasts', '' );
+		if ( ! empty( $ss_podcasting_podmotor_import_podcasts ) ) {
+			return;
+		}
+
+		// check if there is at least one podcast to import
+		$podcast_query = ssp_get_existing_podcasts();
+		if ( $podcast_query->have_posts() ) {
+			add_action( 'admin_notices', array( $this, 'existing_podcasts_notice' ) );
+		}
+	}
+
+	/**
+	 * Show 'existing podcast' notice
+	 */
+	public function existing_podcasts_notice() {
+		$podcast_import_url = add_query_arg( array(
+			'post_type' => $this->token,
+			'page'      => 'podcast_settings',
+			'tab'       => 'import'
+		) );
+		$ignore_message_url = add_query_arg( array( 'podcast_import_action' => 'ignore' ) );
+		$message            = '';
+		$message            .= '<p>You\'ve connected to your Castos account and you have existing podcasts that can be imported.</p>';
+		$message            .= '<p>You can <a href="' . $podcast_import_url . '">import your existing podcasts to Castos.</a></p>';
+		$message            .= '<p>Alternatively you can <a href="' . $ignore_message_url . '">dismiss this message.</a></p>';
+		?>
+		<div class="notice notice-info">
+			<p><?php _e( $message, 'ssp' ); ?></p>
+		</div>
+		<?php
+	}
+
+}
